requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
celery    ==           5.2.7
certifi     ==         2022.12.7
cffi          ==       1.15.1
charset-normalizer==   3.1.0
click              == 8.1.3
click-didyoumean    == 0.3.0
click-plugins      ==  1.1.1
click-repl         ==  0.2.0
colorama           ==  0.4.6
cryptography       ==  40.0.2
Deprecated         ==  1.2.13
Django             ==  4.2.1
django-cors-headers == 3.14.0
django-oauth-toolkit ==2.2.0
djangorestframework  ==3.14.0
idna               ==  3.4
jwcrypto            == 1.4.2
kombu               == 5.2.4
mariadb             == 1.1.6
mysqlclient          ==2.1.1
oauthlib             ==3.2.2
packaging           == 23.1
Pillow              == 9.5.0
pip                 == 22.3.1
prompt-toolkit      == 3.0.38
pycparser           == 2.21
pytz                == 2023.3
requests            == 2.30.0
setuptools          == 65.5.0
six                 == 1.16.0
sqlparse            == 0.4.4
tzdata              == 2023.3
urllib3             == 2.0.2
vine                == 5.0.0
wcwidth             == 0.2.6
wrapt               == 1.15.0
